+++ R/ModuleEigengenes.R
@@ -100,7 +100,9 @@ ComputeModuleEigengene <- function(
     cur_harmony <- harmony::RunHarmony(
       seurat_obj,
       group.by.vars=group.by.vars,
-      reduction.use="ME", verbose=verbose, assay.use=assay, ...
+      reduction.use="ME", 
+      verbose=verbose, 
+      ...
     )@reductions$harmony
     ha <- cur_harmony@cell.embeddings[,pc_dim]
     ha_loadings <- cur_pca@feature.loadings[,pc_dim]